Enhance pattern command with a pause
The pattern command can now process a pause symbol.  The string
representation is adjusted to the way it is used in the `mnomer`
project.

This change resulted in representation change of the pattern type.  The
beat pattern is now its own type BeatPattern with a nested `enum` of
type `char` which leads to a more idiomatic approach.