Fix pause length and add help command
The pause AudioSignal length is adjusted adjusted according to the current
bpm.  Before it was almost non-existant.

A standard help command implementation give all defined commands.  It is
prepared to output the help message of a specific command but the
command type does not yet have a help message as a separate member.